# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos": {
      "post": {
        "description": "Pre-registers the video as a draft in the caller channel, opens a multipart upload in object storage and returns one presigned URL per part. File bytes are sent by the client directly to storage.",
        "operationId": "VideosController_initiateUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created and upload started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start a video upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/upload/part-urls": {
      "post": {
        "description": "Resume path for an interrupted upload: returns a fresh presigned URL for each requested part, so only the missing parts are transferred again.",
        "operationId": "VideosController_reissuePartUrls",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "V1StGXR8Z5j",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPartUrlsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URLs re-issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartUrlsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or a part number is out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this slug in the caller channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The upload is no longer open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Re-issue presigned URLs for parts of an open upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/upload/complete": {
      "post": {
        "description": "Completes the multipart upload with the ETags of every part, verifies the stored size and moves the video to `processing`, where the worker picks it up.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "V1StGXR8Z5j",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Upload completed and queued for processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The parts do not cover the upload exactly",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this slug in the caller channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The upload is no longer open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Storage rejected the parts, or the stored size is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Finish the upload and queue the video for processing",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}": {
      "get": {
        "description": "Returns the video lifecycle status, the metadata extracted during processing, a presigned thumbnail URL once it exists and the failure reason when processing failed.",
        "operationId": "VideosController_getVideo",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "V1StGXR8Z5j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The video",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this slug in the caller channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a video and its processing state",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/stream": {
      "get": {
        "description": "Returns a presigned URL the player requests directly from storage. Range requests are answered by storage with 206 Partial Content — the bytes never pass through the API.",
        "operationId": "VideosController_streamVideo",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "V1StGXR8Z5j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playback URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryUrlResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this slug in the caller channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video is not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a playback URL for the video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{slug}/download": {
      "get": {
        "description": "Returns a presigned URL that serves the original file as an attachment, named after the uploaded file.",
        "operationId": "VideosController_downloadVideo",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "V1StGXR8Z5j",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryUrlResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No video with this slug in the caller channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The video is not ready yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a download URL for the video",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InitiateUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "example": "clip.mp4",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "video/mp4",
              "video/webm"
            ],
            "example": "video/mp4"
          },
          "sizeBytes": {
            "type": "number",
            "example": 209715200,
            "minimum": 1,
            "maximum": 10737418240,
            "description": "Declared file size in bytes; verified after completion."
          },
          "title": {
            "type": "string",
            "example": "My clip",
            "maxLength": 100,
            "description": "Defaults to the file name without its extension."
          }
        },
        "required": [
          "fileName",
          "mimeType",
          "sizeBytes"
        ]
      },
      "UploadPartUrlDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "example": 1
          },
          "url": {
            "type": "string",
            "description": "Presigned URL the client PUTs this part to."
          }
        },
        "required": [
          "partNumber",
          "url"
        ]
      },
      "InitiateUploadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "V1StGXR8Z5j"
          },
          "title": {
            "type": "string",
            "example": "clip"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "error"
            ],
            "example": "draft"
          },
          "partSizeBytes": {
            "type": "number",
            "example": 104857600
          },
          "partCount": {
            "type": "number",
            "example": 2
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPartUrlDto"
            }
          },
          "partUrlsExpireAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "status",
          "partSizeBytes",
          "partCount",
          "parts",
          "partUrlsExpireAt"
        ]
      },
      "RequestPartUrlsDto": {
        "type": "object",
        "properties": {
          "partNumbers": {
            "example": [
              2
            ],
            "description": "Part numbers to re-issue, each between 1 and the video part count.",
            "minItems": 1,
            "maxItems": 10000,
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "partNumbers"
        ]
      },
      "PartUrlsResponseDto": {
        "type": "object",
        "properties": {
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPartUrlDto"
            }
          },
          "partUrlsExpireAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "parts",
          "partUrlsExpireAt"
        ]
      },
      "CompletedPartDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "etag": {
            "type": "string",
            "example": "\"d41d8cd98f00b204e9800998ecf8427e\"",
            "description": "The ETag header storage returned when the part was uploaded."
          }
        },
        "required": [
          "partNumber",
          "etag"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "parts": {
            "description": "Every part of the upload, exactly once.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedPartDto"
            }
          }
        },
        "required": [
          "parts"
        ]
      },
      "CompleteUploadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "V1StGXR8Z5j"
          },
          "status": {
            "type": "string",
            "enum": [
              "processing"
            ],
            "example": "processing"
          }
        },
        "required": [
          "id",
          "slug",
          "status"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "V1StGXR8Z5j"
          },
          "title": {
            "type": "string",
            "example": "My clip"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "error"
            ]
          },
          "mimeType": {
            "type": "string",
            "example": "video/mp4"
          },
          "sizeBytes": {
            "type": "number",
            "example": 209715200
          },
          "durationSeconds": {
            "type": "number",
            "nullable": true,
            "example": 12.5
          },
          "width": {
            "type": "number",
            "nullable": true,
            "example": 1920
          },
          "height": {
            "type": "number",
            "nullable": true,
            "example": 1080
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "description": "Codecs, bitrate, container format and frame rate."
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "Presigned GET URL; null until the thumbnail exists."
          },
          "processingError": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "status",
          "mimeType",
          "sizeBytes",
          "createdAt",
          "updatedAt"
        ]
      },
      "DeliveryUrlResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Presigned GET URL the client requests directly from storage."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "url",
          "expiresAt"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}