# nestjs-project/src/videos/dto/delivery-url-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class DeliveryUrlResponseDto {
  @ApiProperty({
    description: 'Presigned GET URL the client requests directly from storage.',
  })
  url: string;

  @ApiProperty({ format: 'date-time' })
  expiresAt: string;
}