# nestjs-project/src/videos/dto/part-urls-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { UploadPartUrlDto } from './upload-part-url.dto';

export class PartUrlsResponseDto {
  @ApiProperty({ type: [UploadPartUrlDto] })
  parts: UploadPartUrlDto[];

  @ApiProperty({ format: 'date-time' })
  partUrlsExpireAt: string;
}