# nestjs-project/src/videos/dto/upload-part-url.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class UploadPartUrlDto {
  @ApiProperty({ example: 1 })
  partNumber: number;

  @ApiProperty({ description: 'Presigned URL the client PUTs this part to.' })
  url: string;
}